+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Represents the **NuGet** versions.
 
+## v1.0.16
+- *Fixed:* SQL using batch, i.e. `GO` statement, were not being split and executed (indepently) correctly.
+- *Fixed:* Standard pattern of `CancellationToken` added to all `Async` methods.
+
 ## v1.0.15
 - *Fixed:* `Reset` command updated to load embedded SQL resource correctly.